dont bubble up invalid id exceptions [TODO: let bubble up and allow api to handle it]

+++ app/services/list_item_service.py
@@ -1,4 +1,5 @@
 from app import db
+from app.exceptions import InvalidListItemIDException
 from app.models.list_item import ListItem as ListItemModel
 from app.interfaces.list_item import ListItem
 
@@ -17,14 +18,22 @@ class ListItemService:
         return [ListItem(li) for li in list_items]
 
     def get_list_item(self, list_item_id):
-        return ListItem.for_id(list_item_id)
+        try:
+            return ListItem.for_id(list_item_id)
+        except InvalidListItemIDException:
+            return None
 
     def delete_list_item(self, list_item_id):
         list_item = self.get_list_item(list_item_id)
-        list_item.delete()
+
+        if list_item:
+            list_item.delete()
+            return True
+        return False
 
     def update_list_item(self, list_item_id, name=None, description=None, list_id=None):
         list_item = self.get_list_item(list_item_id)
 
-        return list_item.update(name=name, description=description, list_id=list_id)
-
+        if list_item:
+            return list_item.update(name=name, description=description, list_id=list_id)
+        return None
